// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  admin: 'admin',
  editor: 'editor',
  writer: 'writer'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserType = {
  student: 'student',
  teacher: 'teacher'
} as const

export type UserType = (typeof UserType)[keyof typeof UserType]


export const Gender = {
  male: 'male',
  female: 'female',
  others: 'others'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const UserAuditLogType = {
  login: 'login',
  autoCreatePost: 'autoCreatePost',
  writerCreateEntity: 'writerCreateEntity',
  writerEditEntity: 'writerEditEntity',
  editorApproveEntity: 'editorApproveEntity',
  adminApproveEntity: 'adminApproveEntity',
  adminPublishEntity: 'adminPublishEntity',
  unpublishEntity: 'unpublishEntity',
  deleteEntity: 'deleteEntity',
  uploadImage: 'uploadImage',
  deleteImage: 'deleteImage'
} as const

export type UserAuditLogType = (typeof UserAuditLogType)[keyof typeof UserAuditLogType]


export const EntityType = {
  post: 'post',
  page: 'page',
  club: 'club',
  activity: 'activity',
  project: 'project',
  course: 'course',
  faculty: 'faculty'
} as const

export type EntityType = (typeof EntityType)[keyof typeof EntityType]